Composer:install should accept an optional path argument
At the moment, the install task is hard coded to run within
`release_path`. This edit should allow a user to specify a different
execution path, that falls back to `release_path`. This would allow the
use case that I currently have, where I am trying to use
`composer:install` to install the dependencies for a submodule, that is
not located in my root `release_path`.

I say should, because I haven't a lot of experience with rake or
building gems. ;) Hopefully this'll just work, and if not, provoke
discussion.

Fixes #26 (hopefully)