Add optional copy_packages task
Copies the vendor/ dir from the previous release. This can bring a significant reduction in deploy time when there are many dependencies. Note that using this feature makes vendor/ cumulative. That is, as packages are dropped they will not be automatically removed from vendor/ on deploy.

Adapted for Capistrano 3 from http://capifony.org/cookbook/speeding-up-deploy.html