add options for composer and php binaries
Added `:composer_bin` and `:composer_php` options for better control of
which binaries to use.
When not explicitly set, the existence of the `composer` and `php`
executables is now validated before the tasks are run.

Also Composer is now always invoked through php. For this the path to
the `composer` executable is expanded by calling `which composer` on
the remote host.

When `:composer_bin` is set to `:local`, `composer:install_executable`
is automatically invoked on deployment. So no more manual hooking of
additional tasks in the deploy configuration.